refactor: rename mcp/ to skill-market/ and server.ts to index.ts
The `mcp/` name described the transport mechanism (Model Context Protocol),
not what the module actually is — the agent-facing entry into the on-chain
skill marketplace. Readers (and reviewers) kept reading it as "claude/codex MCP
config" rather than "the skill market's gateway", so the name was hiding the
module's purpose.

- src/mcp/           -> src/skill-market/
- skill-market/server.ts -> skill-market/index.ts (it's the folder's entry: a
  tool index + dispatcher over searchSkills/buySkill, not a generic "server")
- index.spec.ts + root index.ts import paths updated to match

Tool/function names (search_skills, buy_skill, createAgentMcpServer) are left
unchanged — they're the MCP-protocol names the engines call, and they're already
clear. 118 tests pass, build clean.
