+++ src/skill-market/skill-shopping.md
